adding config elements num_epochs and eval_steps fixing label type float

+++ utils.py
@@ -10,7 +10,7 @@ def get_file_data(filename):
     data_file = open(filename, 'r', encoding='utf8')
     for line in data_file.readlines()[1:]:
         info = line.split('\t')
-        dataset.append(InputExample(texts=[info[2].strip(), info[3].strip()], label=int(info[1].strip())))
+        dataset.append(InputExample(texts=[info[2].strip(), info[3].strip()], label=float(info[1].strip())))
     data_file.close()
     return dataset
